+++ fluent_contents/plugins/sharedcontent/models.py
@@ -23,6 +23,10 @@ class SharedContent(TranslatableModel):
 
     parent_site = models.ForeignKey(Site, editable=False, default=_get_current_site)
     slug = models.SlugField(_("Template code"), help_text=_("This unique name can be used refer to this content in in templates."))
+    is_cross_site = models.BooleanField(_("Share between all sites"), blank=True, default=False,
+        help_text=_("This allows contents can be shared between multiple sites in this project.<br>\n"
+                    "Make sure that any URLs in the content work with all sites where the content is displayed."))
+
     contents = PlaceholderField("shared_content", verbose_name=_("Contents"))
 
     # NOTE: settings such as "template_name", and which plugins are allowed can be added later.